cangjie5: Check the characters classified as Simplified only against a Traditional dictionary
See: https://github.com/mike-fabian/ibus-table/issues/100#issuecomment-1019993348

For example, if I search for that 冯 character, this is the page I
get:

https://dict.revised.moe.edu.tw/search.jsp?md=1&word=%E5%86%AF#searchL

In the page there is the phrase: 查無資料 (which means 'not found')

So if that script returns a page with this phrase, that means the
character isn't used in Trad. If no such phrase is present on the
page, then it's used.